Split bot.py into modular structure: shared, tasks, utils, and command handlers
