+++ app/discord_bot/config.py
@@ -1,0 +1,4 @@
+import os
+
+API_URL = os.getenv("API_URL", "http://localhost:8002")
+DISCORD_TOKEN = os.getenv("DISCORD_TOKEN")
